+++ src/build-packages.ts
@@ -1,6 +1,7 @@
 import fs from 'fs-extra';
 import path from 'path';
 import { build } from 'rocket-punch';
+import { buildMessageHandler } from 'rocket-punch/message-handlers/build';
 
 [path.join(process.cwd(), 'node_modules/@ssen'), path.join(process.cwd(), 'node_modules/rocket-punch')].forEach((p) => {
   if (fs.existsSync(p)) {
@@ -15,4 +16,5 @@ import { build } from 'rocket-punch';
 build({
   cwd: process.cwd(),
   dist: path.join(process.cwd(), 'dist'),
+  onMessage: buildMessageHandler,
 });
