+++ src/publish-packages-e2e.ts
@@ -1,5 +1,6 @@
 import path from 'path';
 import { publish } from 'rocket-punch';
+import { publishMessageHandler } from 'rocket-punch/message-handlers/publish';
 import yargs from 'yargs';
 
 const argv = yargs.argv;
@@ -10,4 +11,5 @@ publish({
   skipSelection: true,
   tag: argv.tag as string,
   registry: argv.registry as string,
+  onMessage: publishMessageHandler,
 });
